Tweak: Use `IconControl` component
