+++ package.json
@@ -23,7 +23,7 @@
 		"@dnd-kit/sortable": "^8.0.0",
 		"@dnd-kit/utilities": "^3.2.2",
 		"@edge22/block-styles": "^1.1.18",
-		"@edge22/components": "^1.1.6",
+		"@edge22/components": "^1.1.9",
 		"@edge22/styles-builder": "^1.1.37",
 		"@wordpress/block-editor": "^12.12.0",
 		"@wordpress/blocks": "^12.21.0",
