--- src/components/icon-control/editor.scss
@@ -1,24 +1,0 @@
-.gb-icon-control {
-	.components-base-control {
-		margin-bottom: 0;
-	}
-
-	&__preview {
-		display: flex;
-		align-items: center;
-		gap: 10px;
-		margin-top: 5px;
-
-		& > span {
-			line-height: 0;
-			display: block;
-			max-width: 50%;
-		}
-
-		svg {
-			height: 20px;
-			width: auto;
-			max-width: 100%;
-		}
-	}
-}
