--- src/components/icon-control/index.js
@@ -1,4 +1,0 @@
-import { IconControl } from './IconControl';
-import { IconModal } from './IconModal';
-
-export { IconControl, IconModal };
