--- src/components/icon-library/index.js
@@ -1,3 +1,0 @@
-import { IconLibrary } from './IconLibrary.jsx';
-
-export { IconLibrary };
